+++ src/app.js
@@ -9,5 +9,4 @@ app.use(router);
 
 app.listen(process.env.PORT, () => {
     console.log("Magic happens on port " + process.env.PORT);
-  });
-  
+});
